Moved FileSessionHandler to Handler folder
